fix: add pull request remarks
